Fix generate JSON Schema for arrays, when having different items
In generated JSON Schema for arrays, `items` was an object before.
Now it is an array, which resolves the problem of having different
kind objects in the array. The `properties` are added for each of those,
meaning each of them can have a different schema.

Note that JSON Schema `additionalItems` is `true` by default.
This had not effect before as `items` was an object. For more info, see:
http://json-schema.org/draft-04/json-schema-validation.html#rfc.section.5.3.1