Release 1.0.0rc2, with keyword `Clear Expectations` put back
The schemas are possible to organize so that using this keyword is not
necessary; I will explain my reasoning on this in the tutorial via
examples. It took a few hours of thinking for me as well to realize this.

That secondary though if something is broken.

If this keyword is to be deprecated later, 1.1.0 is the earliest and
we will then have a proper warning of this at 1.0.0.